Use pkg-config to find libxml2 and libcurl on Linux [SR-5608]
This allows linking swift-corelibs-foundation to packages with flags
defined by `pkg-config`, rather than hardcoded ones. There are a couple
of structural changes as well, namely that `pkg-config` will now be
executed during `configure` rather than later by `ninja`. This seems
acceptable as it's how cmake and autoconf work, but means a reconfigure
will be required if any transitive dependency of a package we depend on
files changes.

The other structural change is that `-lcurl` will be passed
unconditionally on platforms without pkg-config support (technically the
`pkg_config_required` variable set). This makes curl a harder dependency
on those platforms, but that seems unavoidable given that it's the only
implementation of `URLSession`.