Reduce test compile times.
I noticed that the tests were taking quite a while to compile.

Using the "-warn-long-expression-type-checking" flag, I measured the
compile time of the test project. The total of types over 100ms was
around 27 seconds before this change and around 25 seconds after (on my
build configuration), so it shaves off a few seconds there.

I also took the opportunity to reduce the number of attempted threads in
the URL test from 640 to 10. There isn't much value add in going above
the number of cores we're going to find on most machines, and that
dramatically reduces the run time of that test.