Fix String.deletingPathExtension Inconsistency
The following code behaves differently on macOS and Linux:

```swift
NSString(string: "..").deletingPathExtension
```

On Linux, the resulting value is `.` while on macOS, it's `..`.

I think the latter make far more sense since `..` mean parent directly,
treating the last dot as "extension" is a bit of a stretch. Plus, this
logic existed on macOS longer (therefore has more existing users). So it
make sense to align Linux Foundation's behavior to that of macOS.