Create a scaffold implementation of FileHandle#waitForDataInBackgroundAndNotify
This change alone is still insufficient. It needs a few details to be
addressed.

1. From the docs it says the function must be called from a thread that
has an active run loop. How does one go about checking if a thread has
an active run-loop? Would that be an assert or precondition or something
else entirely?

2. From the docs it says the function will post the
NSFileHandleDataAvailable notification on the current thread. Presumably
that means the thread that was used to call
waitForDataInBackgroundAndNotify in the first place. How do you specify
the thread when posting a notification?

3. Tests. Are there any suggestions for tests?