Fixed `addObserver`'s obsolescence
As per [this post](https://forums.swift.org/t/found-a-bug-in-foundations-obsolescence-mechanism-in-linux/7730) in the forums.

Fixed with Xiaodi's suggestion.
It passed `ninja test` in my computer and it fixed the problem I described after I rebuilt my binaries with this patch.

I think we need to run a `ninja test` in CI, just to be sure. After that, it should be ready for merge.