closure: make it build on Windows x86_64
Windows x86_64 is a LLP64 environment.  `unsigned long` is not `size_t`.
Adjust the parameter to be the correct size for the allocator.  This is
compatible with the previous implementation on LP64 environments.

Prefer to use the `__sync_bool_compare_and_swap` over the MSVC
intrinsic which is overly strong.  While in the area, remove an unused
compare and swap method.