[SR-7112] Fixed a case where CFStringGetCStringPtr returns a string without null termination
Changed _fastCStringContents (used in CFStringGetCStringPtr) to return NULL if required to return a string containing null termination.
Since this fix, if you call CFStringGetCStringPtr with a Swift-based CFString, it returns NULL.

Before the change: _fastCStringContents always returned a pointer to an ASCII string containing no NULL termination.
This is because the code was using unsafeBitCast to get internal pointer of String of Swift. Swift's string does not guarantee null termination.