Fix argument order for CFArraySetValueAtIndex.
The Swift dispatching of CFArraySetValueAtIndex had the arguments for
value and index switched around. If someone creates a Swift NSArray
which is eventually bridged to CFArray, and some other piece calls
CFArraySetValueAtIndex, CF_SWIFT_FUNCDISPATCHV will delegate to the
Swift method, but will send a pointer where CFIndex is expected, and a
CFIndex where a pointer is expected, probably causing a segmentation
fault.

The test tries to simplify the scenario (not easy to create the right
conditions) and uses unsafeBitCast (which is the same that happens in
the internal _cfObject). The test crashes before the patch is applied.