Fix 32-bit Variable Offset Issues with CoreFoundation
_CFInfo Assumes a 64-bit _cfinfoa type, when it is really 32-bit on 32-bit systems.
_NSCFType assumes a 32-bit _cfinfoa too, making things even more complicated.

There's a couple approaches to addressing this:
1) Back _CFInfo with a UInt that, on paper, matches size with __CFInfoType
2) Add padding so we can guarantee that _CFInfo can always be 64-bit.

Which one you pick depends on if this is guaranteed to be true: UInt == __CFInfoType

Because Android has some interesting macros in the code, this change picks option 2, even though it
does mean an extra 4 bytes for each bridgable object.

In either case, it makes sense that _NSCFType should be backed by a _CFInfo instead
of a raw int type. Share types when possible so we have consistent, understandable behavior.