Add back unnecessarily redacted headers that now exist
I think this might have been an artifact of the old CFLite redactions?
Either way, these headers are now available in this repository. The real
CoreFoundation guards these includes with `#ifndef CF_OPEN_SOURCE` but
that doesn't seem relevant here. There are also other includes guarded
by that which are not (yet? 😆) available in this repository, so I did
not include them here.