build: adjust for split libdispatch SDK overlay
This adjust the build to explicitly link against swiftDispatch to enable
the split SDK overlay on non-Darwin targets.  This would be fixed when
we switch over to the CMake based build which uses the swift driver to
link which will use the auto-link-extract functionality to make the ELF
targets build.