CoreFoundation: use #include instead of #import
`#import` when building for Windows will be treated as the import of a
type library rather than the semantics of Objective-C style `#import`
which does an include once type of operation.  Convert to the `#include`
syntax instead to allow this to be built on Windows as well (e.g. via
clang-cl).