POSIX: guard the inclusion of pthread.h
Not all deployment targets use POSIX threads (pthreads).  Use the POSIX
mandated macro `_POSIX_THREADS` to detect whether POSIX threads are in
use.  Conditionally include `unistd.h` which is the Unix standard header
which provides the definition for `_POSIX_THREADS` if the platform uses
POSIX threads.  Unfortunately, clang does not define `__unix__` for
macOS (Bob Wilson does not recall why this was the case, and we have yet
to test if we can get away with defining `__unix__` for the platform
without breaking existing software).