Base, Locale, PlugIn, RunLoop: spell pthread as CF
Use `CFLock_t` instead of `pthread_mutex_t`.  This has a small semantic
change on Linux where we now use pthreads rather than a cmpxchg to
implement locking.  This makes the behaviour of POSIX-y systems and
Darwin more uniform.  Additionally, we now generically target pthreads
on any system that supports POSIX threading.  Windows continues to
behave differently (though should replace the cmpxchg in favour of
SWRLOCK).

This leaves the open question of why do we not use `pthread_spinlock_t`
to implement the libkern `OSSpinLock*` family of functions.  If we do
so, we should be able to fold away the differences on POSIX and Darwin
systems.