Consolidate building setup to CMake
This adds a new utils/create-build-files script that can be used, via cmake, to produce build setups meant to replace build.py and the Xcode workspace and projects to build swift-corelibs-foundation without building the rest of the Swift codebase.

This isn’t fully done, but allows the Foundation and CoreFoundation targets to build in a generated Xcode (utils/create-build-files --xcode) or Ninja setup (utils/create-build-files --ninja).