CoreFoundation: remove some UID interface on Win32
The UID concept does not map to Windows very well where users are
identified by a GUID and do not have a primary group.  The user
permission model is far richer and cannot be easily distilled to the
simplistic Unix model.  These are not currently used in
swift-corelibs-foundation, so simply elide the interfaces for now.