build: we only support the shared CRT on Win32
Adjust the CFLAGS to indicate that the shared CRT is used since we do
not support the static CRT as that makes the DLL unusable with other
executables as the shared and static CRT cannot be mixed in a single
address space.  The flag itself does not collide on other targets (at
least currently).  If necessary, that can be made target dependent.