Foundation: fix a couple of compile errors
FoundationErrors does not depend on NSSwiftRuntime and thus does not
transitively get the WinSDK import when not building in whole program
mode.  Add an import to get a definition of `DWORD`.

Remoe the `private(set)` from the computed `processIdentifier` property
which is already readonly due to the definition of the computed
property.