build: ensure that we build CoreFoundation for re-export
We wish to re-export CoreFoundation from Foundation.  In order to
accomplish this, the objects need to be built as if they are being built
for a DLL as they will be subsumed into the Foundation DLL.  We do not
necessarily wish to do this if we are building Foundation static
however.  Doing this ensures that we can re-export the symbols from
CoreFoundation on Windows.