build: default to hidden visibility
This will ensure that the internal implementation details of
CoreFoundation do not leak out from the generated library.

A static library should not expose its interfaces for exporting as the
objects will be incorporated into another library or executable which
should not masquerade as a different library to prevent symbol
collisions from multiple libraries statically linking to the same
library.

When building CoreFoundation as a shared library, CoreFoundation_EXPORTS
should be defined (which CMake does by default.  This allows the library
to expose its public interfaces when built in a shared mode.

This work will now enable [Swift]Foundation to hide the internal
implementation detail of CoreFoundation from users.

In order to prevent breakage of existing applications, for now, export
public interfaces always.