FileHandle: make FileHandle build on Windows again
A SOCKET is not a file descriptor on Windows which means that the shared
implementation does not work for Windows.  This allows us to build
again.  It should be possible to provide a special constructor that
takes a SOCKET rather than a fd and wait on that though, so for now, use
NSUnimplemented to guard that path.