build: use `add_subdirectory` instead of `add_external_project`
This switches the CoreFoundation sub-build to use `add_subdirectory`.
This has a few **MAJOR** benefits:
- CoreFoundation changes get tracked and cause compile/link triggers
- CoreFoundation warnings are surfaced in the build
- CoreFoundation dependencies are reflected in the right location
- Foundation dependencies are reflected in the right location
- The overall build is simpler

Eventually, it should be possible to further simplify this.

Unfortunately, a couple of hacks are needed to make this work.  The
sub-projects properties must be cached, but they were just passthrough
previously, so this is not too terrible.  Additionally, the library
control know needs to be saved/restored around it.  The worst thing here
is the need to manually collect the dependency as we do not have a
proper library target for the Foundation (Swift) library.  When CMake
gains proper Swift support, it should be possible to use `add_library`
and get a property target that we can use `target_link_libraries` with
permitting the dependency tracking to allow us to get that without
manual intervention.