Tools: add a XSL to create the Windows/Olson mappings
Since Windows uses its own names for timezones, the Unicode corsotium provides
the canonical mapping for the timezones.  Use the Unicode CLDR to generate the
mapping plist requires transformation, which this XSLT provides.  These XSLs
allow the creation of both the Olson -> Windows mapping and the Windows -> Olson
mapping.  Since the latter only preserves the primary name and not the secondary
ones, we split the inverted database.