Foundation: initialize CFSocket on startup
Windows requires the networking subsystem (WinSock) to be initialized.
However, the user expects to be able to use the networking APIs in
Foundation without performing any one-time initialization.  Furthermore,
due to the macro usage in the initialization path, it is easier to
perform the initialization in C than in Swift.  CoreFoundation has a SPI
for this (`__CFSocketInitializeWinSock`).