disable test_currencyPluralMinimumIntegerDigits
test_currencyPluralMinimumIntegerDigits does not and did never work on
Ubuntu 14.04. In Swift 5 it works because a newer ICU is vendored which
fixes the issue. But in Swift 4.2.x we don't vendor ICU so we can't get
this test to work on Ubuntu 14.04, hence removing as it's no worse than
Swift 4.2.3.