disable test_currencyAccountingMinimumIntegerDigits & test_currencyISOCodeMinimumIntegerDigits
test_currencyAccountingMinimumIntegerDigits does not work with Ubuntu
14.04's libICU and therefore we can't run it in CI.

This will be fixed with Swift 5 because Swift 5 vendors libICU which is
the right solution but too big of a change for Swift 4.2.

test_currencyISOCodeMinimumIntegerDigits has to be disabled too even
though the code does not crash on Swift 4.2.3 on Ubuntu 14.04. The
reason is that commit 3290d07b3c1a34214ca0d0b15ab063c4f0c52946 corrects
the NumberFormatting styles. Previously, currencyISOCode used to map to
`7` which is incorrect. Now it correctly maps to `8` but unfortunately
that is unsupported by 14.04's libICU...