Parity: NSCoder miscellanea
- Added missing API: `DecodingFailurePolicy.raiseException`. It is now the default like on Darwin.

- Match Darwin behavior while decoding if you decode without raising exceptions.

- `decodeTopLevel…` variants are noted unavailable on swift-corelibs-foundation, with rationale and replacement directions.

Swift doesn’t have exceptions, and it used to define `.setErrorAndReturn` as the defaults. However, until very recently `failWithError(_:)` was unimplemented and trapped with a fatal error, effectively meaning that all coding failures were behaving as if `.raiseException` was set in a Darwin Swift-only app: by crashing.

Since code written to 4.x-5.0 may be relying on this behavior, we make it explicit and consistent with Darwin: there is now a `.raiseException` constant, it’s the default, and it fails with a fatal error in `failWithError(_:)`. The new secure initializers (at https://github.com/apple/swift-corelibs-foundation/pull/2102) set `.setErrorAndReturn` just like on Darwin, and thus won’t raise and instead just thrown an error.

Tests will be added.