[android] Fix some parts of Bundle in Android.
While Android does not follow FHS, and it doesn't seem to have something
similar published, it follows some conventions of Unix. One of them
is that libraries are normally named prefixed with "lib" and suffixed
with ".so" like Linux. Executables are also normally named without any
prefix nor suffix.

The changes in this commit lift up restrictions that were in place for
Android, and allow some of the Bundle tests to pass on Android (and lift
another restriction in TestObjCRuntime).

While bundles might never be a common construct in Android, these
changes at least allow to test some of the scenarios, in case someone
needs this behaviour.