Fix StringTransform usage of forced casts.
In Linux, a (constant?) CFString cannot be force casted to String
directly. The code generated for those values with optimizations enabled
was simply a crash. The changes uses `_swiftObject` to correctly
transform to a `Swift.String`.

There some other cases in Foundation where other enums are initialized
from CF values, but in those cases the literals of the CF values are
used, which avoids this problem.