[android] Fix tilde tests in Android.
In Android, the home directory is the root directory. Foundation returns
the right one, but the tests where checking against having a final
trailing slash, which is not true if the home directory is the root
directory.

The change adds an extra check to avoid the failure in case the
resulting tilde expansion is just the root directory. The comment
explains why the change is needed.

Additionally, instead of using `+` to add two strings,
`appendingPathComponent` is used to build an expected value from the
result of `NSHomeDirectory()`, in order to deal with cases where the
`NSHomeDirectory()` has a trailing slash.