Foundation: wrap `_withStackOrHeapBuffer`
`_withStackOrHeapBuffer` is used in multiple functions in Data which are
marked as `@inlineable`.  Since `_withStackOrHeapBuffer` is inlined into
the body, this results in the users getting references to
`_CFThreadIsMain` which is SPI in CoreFoundation for Foundation.
Furthermore, since the CoreFoundation interfaces are not exported from
Foundation, we cannot fulfill the dependency.  Create a wrapper over
`_withStackOrHeapBuffer` to allow the dependency to be contained.