build: replicate `/M{T,D}{,d}` flag support
Windows requires explicitly setting up the C library that you link
against and the flags for building against it as they share the same set
of headers.  Replicate the behaviour of the C frontend for the build to
ensure that we link against the symbols properly and use the correct
library at link time.