Foundation/CoreFoundation: repair bridging for `NSLocale`
`NSLocale` is toll-free bridged to `CFLocale`.  The types must match.
On many targets, the backing store for `_lock` was being elided as the
value was conditionally defined.

Extract the locking macros/types into a `CFLocking.h` and then include
that into `CFInternal.h` and `ForSwiftFoundationOnly.h`.  This allows
the underlying type information to be shared into Swift permitting the
`__CFLocale` type to be reconstructed without guessing as to the
underlying type.

This repairs the bridging of `NSLocale` on Windows and allows us to get
further in the test suite.