[android] Enable Foundation.Host
Depends on apple/swift#24286

The implementation of Host was disabled for Android, probably because
Glibc didn't include one header (netdb.h) which was needed to compile
this file. But changing Glibc is not enough, because the Android
implementation of that function and other bits differ slightly from
Linux implementation. For those cases private overrides are provide, to
keep the code unchanged.

There's also a little bit of formatting in one of Windows parts, which
was missing two spaces of indentation (probably checking without
whitespace enabled is good for this diff).