[android] Enable Thread tests were possible.
Enable most of the Thread test for Android. The only tests that cannot
pass right now are the ones related to backtraces, which do not use the
Linux API backtrace and would need an specific implementation.

And while Android did not get pthread_getname_np until API 26, the low
level prctl(PR_GET_NAME) works with the same restriction as in Linux
(only 15 characters). With that change, the test about thread names also
passes in Android.