Avoid calling stat twice for each item returned by NSURLDirectoryEnumerator
fts_read calls stat internally to determine the type of each file that is visited. If URL(fileURLWithPath:) is called
with a path that doesn't end with a slash, it calls stat to determine whether the path represents a file or a
directory. We know from the FTSENT's fts_info field whether the entry represents a file or a directory, so we can
instead call URL(fileURLWithPath:isDirectory:) and avoid the second stat call.

This reduces the runtime of FileManager.enumerator(at:includingPropertiesForKeys:options:errorHandler:) by around 40%.