[gardening] Remove conditional casts from Error to NSError.
Since Swift 3 (IIRC) Error can always be casted to NSError, so the
conditional casts are always going to succeed and they produce warnings.
However because in macOS SwiftFoundation.NSError is not the ObjC
NSError, the compiler misses this, and might fail to mark some catch as
exhaustives.

To avoid one warning per use of NSError, move all the warnings into a
single one in a utility function, used from all the places when Error
needs to be transformed into NSError. In a couple of cases, the code is
simplified because a branch disappears.

This should remove a lot of warnings while compiling the tests.