[gardening] Fix assign to a weak variable releasing immediately.
The compiler was warning about this assign to a weak variable, where the
value was going to be released immediately. Modify the code to hold the
value in a strong variable during the block to avoid the warning (and
probably a failing test).

This test is actually disabled, so nothing should change.